[browser] Move boot config tests to Wasm.Build.Tests
Move GenerateWasmBootJson test coverage from the temporary tasks.tests project into Wasm.Build.Tests, remove the obsolete tasks.tests csproj, and tag the moved tests as no-workload.


Copilot-Session: 2db6a9dd-0c66-4a48-b2f6-36c7329a5e0e
